requirements.txt
python-telegram-bot==21.4
playwright==1.49.0
